+++ locale/init.lua
@@ -2,11 +2,10 @@ local i18n = require 'i18n'
 local path = require 'lib.path'
 
 local LOCALE_ROOTFP = path.prjoin('locale')
-local SEP = package.config:sub(1, 1)
 local locales = {en = true, pt = true}
 
 for locale in pairs(locales) do
-  i18n.loadFile(LOCALE_ROOTFP .. SEP .. locale .. '.lua')
+  i18n.loadFile(path.join(LOCALE_ROOTFP, locale .. '.lua'))
 end
 i18n.setFallbackLocale('en')