+++ make-locale.lua
@@ -1,0 +1,16 @@
+return {
+  en = {
+    logs = {err = 'ERROR', ok = 'OK'},
+    missing_command = 'please specify `build`, `install`, `config` or `fonts`',
+    build = {
+      ok = 'YGOFabrica has been succesfully built!'
+    }
+  },
+  pt = {
+    logs = {err = 'ERRO', ok = 'OK'},
+    missing_command = 'por favor especifique `build`, `install`, `config` ou `fonts`',
+    build = {
+      ok = 'YGOFabrica foi compilada com sucesso!'
+    }
+  }
+}